Runtime changes up to the point of resolution in getCallInfo
I have modified the behavior based on the principle that the
virtual static methods shouldn't go in the vtable. For this purpose
they are treated as 'non-virtual' when building the interface
slot table. I'm now trying to figure out how to traverse the
MethodImpls in MethodTable during the constrained lookup, I haven't
yet managed to identify whether they are already accessible via
the existing MethodTable API or whether it will require adding
a new table.

Thanks

Tomas